feat: implement JpegValidationResult class for comprehensive JPEG validation results
