+++ Graphics/Rasters/src/Root/Jpegs/JpegValidator.cs
@@ -13,25 +13,41 @@ public static class JpegValidator
 		ArgumentNullException.ThrowIfNull(jpeg);
 
 		var result = new JpegValidationResult();
-		
-		// Validate dimensions
+
+		ValidateDimensions(jpeg, result);
+		ValidateQuality(jpeg, result);
+		ValidateColorModeAndSamples(jpeg, result);
+		ValidateBitsPerSample(jpeg, result);
+		ValidateCompressionRatio(jpeg, result);
+		ValidateEncodingConstraints(jpeg, result);
+		ValidateChromaSubsampling(jpeg, result);
+
+		return result;
+	}
+
+	private static void ValidateDimensions(IJpegRaster jpeg, JpegValidationResult result)
+	{
 		if (jpeg.Width <= 0)
 			result.AddError($"Invalid width: {jpeg.Width}. Width must be greater than 0.");
-		
+
 		if (jpeg.Height <= 0)
 			result.AddError($"Invalid height: {jpeg.Height}. Height must be greater than 0.");
-		
+
 		if (jpeg.Width > JpegConstants.MaxDimension)
 			result.AddError($"Width exceeds maximum: {jpeg.Width} > {JpegConstants.MaxDimension}.");
-		
+
 		if (jpeg.Height > JpegConstants.MaxDimension)
 			result.AddError($"Height exceeds maximum: {jpeg.Height} > {JpegConstants.MaxDimension}.");
+	}
 
-		// Validate quality
+	private static void ValidateQuality(IJpegRaster jpeg, JpegValidationResult result)
+	{
 		if (jpeg.Quality < JpegConstants.MinQuality || jpeg.Quality > JpegConstants.MaxQuality)
 			result.AddError($"Invalid quality: {jpeg.Quality}. Quality must be between {JpegConstants.MinQuality} and {JpegConstants.MaxQuality}.");
+	}
 
-		// Validate color mode and samples per pixel
+	private static void ValidateColorModeAndSamples(IJpegRaster jpeg, JpegValidationResult result)
+	{
 		var expectedSamples = jpeg.ColorMode switch
 		{
 			JpegColorMode.Grayscale => 1,
@@ -45,24 +61,30 @@ public static class JpegValidator
 			result.AddError($"Invalid color mode: {jpeg.ColorMode}.");
 		else if (jpeg.SamplesPerPixel != expectedSamples)
 			result.AddError($"Invalid samples per pixel: {jpeg.SamplesPerPixel}. Expected {expectedSamples} for {jpeg.ColorMode} color mode.");
+	}
 
-		// Validate bits per sample
+	private static void ValidateBitsPerSample(IJpegRaster jpeg, JpegValidationResult result)
+	{
 		if (jpeg.BitsPerSample != JpegConstants.BitsPerSample)
 			result.AddError($"Invalid bits per sample: {jpeg.BitsPerSample}. JPEG supports only {JpegConstants.BitsPerSample} bits per sample.");
+	}
 
-		// Validate compression ratio
+	private static void ValidateCompressionRatio(IJpegRaster jpeg, JpegValidationResult result)
+	{
 		if (jpeg.CompressionRatio <= 0)
 			result.AddError($"Invalid compression ratio: {jpeg.CompressionRatio}. Compression ratio must be greater than 0.");
+	}
 
-		// Validate encoding-specific constraints
+	private static void ValidateEncodingConstraints(IJpegRaster jpeg, JpegValidationResult result)
+	{
 		if (jpeg.Encoding == JpegEncoding.Jpeg2000)
 			result.AddWarning("JPEG 2000 format has limited support in many applications.");
+	}
 
-		// Validate chroma subsampling for color modes
+	private static void ValidateChromaSubsampling(IJpegRaster jpeg, JpegValidationResult result)
+	{
 		if (jpeg.ColorMode == JpegColorMode.Grayscale && jpeg.ChromaSubsampling != JpegChromaSubsampling.None)
 			result.AddWarning("Chroma subsampling is not applicable for grayscale images.");
-
-		return result;
 	}
 
 	/// <summary>Validates JPEG file signature.</summary>
@@ -71,7 +93,7 @@ public static class JpegValidator
 	public static bool IsValidJpegSignature(ReadOnlySpan<byte> data)
 	{
 		if (data.Length < 2) return false;
-		
+
 		// Check for SOI marker (Start of Image)
 		return data[0] == 0xFF && data[1] == 0xD8;
 	}
@@ -116,53 +138,5 @@ public static class JpegValidator
 			result.AddError($"Invalid focal length: {metadata.FocalLength}. Focal length must be greater than 0.");
 
 		return result;
-	}
-}
-
-/// <summary>Represents the result of JPEG validation.</summary>
-public class JpegValidationResult
-{
-	/// <summary>Gets a value indicating whether the validation passed.</summary>
-	public bool IsValid => Errors.Count == 0;
-
-	/// <summary>Gets the list of validation errors.</summary>
-	public List<string> Errors { get; } = new();
-
-	/// <summary>Gets the list of validation warnings.</summary>
-	public List<string> Warnings { get; } = new();
-
-	/// <summary>Adds an error to the validation result.</summary>
-	/// <param name="error">The error message to add.</param>
-	public void AddError(string error)
-	{
-		Errors.Add(error);
-	}
-
-	/// <summary>Adds a warning to the validation result.</summary>
-	/// <param name="warning">The warning message to add.</param>
-	public void AddWarning(string warning)
-	{
-		Warnings.Add(warning);
-	}
-
-	/// <summary>Gets a summary of all validation issues.</summary>
-	/// <returns>A formatted string containing all errors and warnings.</returns>
-	public string GetSummary()
-	{
-		var summary = new List<string>();
-		
-		if (Errors.Count > 0)
-		{
-			summary.Add($"Errors ({Errors.Count}):");
-			summary.AddRange(Errors.Select(e => $"  - {e}"));
-		}
-		
-		if (Warnings.Count > 0)
-		{
-			summary.Add($"Warnings ({Warnings.Count}):");
-			summary.AddRange(Warnings.Select(w => $"  - {w}"));
-		}
-		
-		return summary.Count > 0 ? string.Join(Environment.NewLine, summary) : "No validation issues found.";
 	}
 }
